backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase>=2.11.0
pandas==2.2.0
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.9
openai==1.12.0
prophet==1.1.5
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
python-multipart==0.0.9
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx>=0.27.0
httpcore>=1.0.8
websockets>=13.1
email-validator>=2.0.0
